+++ app/scripts/lib/config-manager.js
@@ -6,6 +6,8 @@ const TESTNET_RPC = MetamaskConfig.network.testnet
 const MAINNET_RPC = MetamaskConfig.network.mainnet
 const MORDEN_RPC = MetamaskConfig.network.morden
 const KOVAN_RPC = MetamaskConfig.network.kovan
+const RINKEBY_RPC = MetamaskConfig.network.rinkeby
+
 
 /* The config-manager is a convenience object
  * wrapping a pojo-migrator.
@@ -153,6 +155,9 @@ ConfigManager.prototype.getCurrentRpcAddress = function () {
 
     case 'kovan':
       return KOVAN_RPC
+    
+    case 'rinkeby':
+      return RINKEBY_RPC
 
     default:
       return provider && provider.rpcTarget ? provider.rpcTarget : TESTNET_RPC