+++ ui/lib/account-link.js
@@ -11,6 +11,9 @@ module.exports = function (address, network) {
     case 3: // ropsten test net
       link = `http://ropsten.etherscan.io/address/${address}`
       break
+    case 4: // rinkeby test net
+      link = `http://rinkeby.etherscan.io/address/${address}`
+      break
     case 42: // kovan test net
       link = `http://kovan.etherscan.io/address/${address}`
       break