+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix occasional nonce tracking issue.
 - Fix bug where some events would not be emitted by web3.
 - Fix bug where an error would be thrown when composing signatures for networks with large ID values.
+- Add Rinkeby Test Network to our network list.
 
 ## 3.5.3 2017-4-24